services/docreader/requirements.txt
grpcio
grpcio-tools
protobuf
python-docx
PyPDF2
requests
Pillow
beautifulsoup4
lxml
playwright
asyncio
urllib3
markdownify
mistletoe
goose3[all]
# paddlepaddle==3.0.0  (普通CPU版本已被注释)
paddleocr==3.0.0
markdown
pypdf
cos-python-sdk-v5
textract
antiword
openai

--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu126/
paddlepaddle-gpu==3.0.0 